Extract object store client instantiation
